+++ CHANGELOG.md
@@ -1,3 +1,23 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial Release.
+* **SlideText**: Slides text in from left to right or right to left.
+* **TypewriterText**: Simulates a typewriter effect, revealing text character by character.
+* **WaveMotionText**: Applies a sinusoidal wave motion to each character.
+* **WavingGradientText**: Applies a waving gradient effect to the text.
+* **ShimmerText**: Applies a shimmer shine effect to the text.
+* **FadeText**: Fades text in/out.
+* **ScaleText**: Scales text up and down.
+* **RotationText**: Rotates text back and forth.
+* **MarqueeText**: Moves text horizontally like a marquee.
+* **FallingText**: Makes each character fall into place.
+* **ColorizedText**: Animates text through multiple colors.
+
+## 0.0.2
+
+- Removed unnecessary animation start code.
+
+## 1.0.0
+
+* **BouncingText**: Animates each character with a bouncing effect from left to right or right to left.
+* **ShadowText**: Animates text shadow effect.